refactor of keystone examples
This commit refactors the keystone examples file so
that it is easier to consume from the
openstack_dev_env for testing.

Instead of keeping swift examples in this top level
testing project:

https://github.com/puppetlabs/puppetlabs-openstack_dev_env

instead, it imports the example manifest from this
modules:

  # manifests/site.pp
  import '/etc/puppet/modules-0/swift/examples/site.pp'

This is intended to solve two problems:
1. previously, the example code was ported to
this project, this caused the example code that
ships with each module to get out of date (since
testing is only performed on the dev_env)
2. prevents me from having to maintain multiple
examples.

There are a few things that had to be refactored for
this to work:

1. keystone example is moved to examples/keystone.pp

The dev_env already has a definition for keystone

2. example needed to be updated to use hiera

so that the dev_env can override variables and
control the config of swift through the dev env.